apps/home_page/serializers.py
from rest_framework import serializers

from apps.home_page.models import Review


class ReviewSerializer(serializers.ModelSerializer):
    class Meta:
        model = Review
        fields = ['title', 'text']
