requirements.txt
asgiref==3.6.0
Django==4.1.6
django-environ==0.9.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
environ==1.0
importlib-metadata==6.0.0
Markdown==3.4.1
Pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.12.0
